Remove RequireJS, jQuery, and Underscore for straight Babel & Webpack
