+++ demo/index.html
@@ -1,9 +1,11 @@
-<!-- index.html -->
-<html>
+<!DOCTYPE html>
+<html lang="en">
   <head>
     <title>FlowTip.React Demo</title>
-    <script data-main="main" src="require-2.1.17.min.js"></script>
-    <link rel="stylesheet" href="index.css" />
+    <link rel="stylesheet" href="/index.css" />
   </head>
   <body></body>
+  <script src="/demo.js"></script>
+</body>
 </html>
+
