--- demo/noamd.html
@@ -1,16 +1,0 @@
-<!-- index.html -->
-<html>
-  <head>
-    <title>FlowTip.React Demo</title>
-    <script src="../node_modules/react/dist/react.js"></script>
-    <script src="../node_modules/react/dist/JSXTransformer.js"></script>
-    <script src="../node_modules/jquery/dist/jquery.js"></script>
-    <script src="../node_modules/underscore/underscore.js"></script>
-
-    <script src="../flowtip-noamd.min.js"></script>
-    <script type="text/jsx" src="noamd.jsx"></script>
-
-    <link rel="stylesheet" href="index.css" />
-  </head>
-  <body></body>
-</html>
