+++ package.json
@@ -1,8 +1,8 @@
 {
-  "name": "FlowTip",
+  "name": "flowtip",
   "version": "2.0.0",
   "description": "A flexible, adaptable and easy to use tooltip positioning library.",
-  "main": "flowtip.js",
+  "main": "build/flowtip.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/qiushihe/flowtip.git"
@@ -19,17 +19,28 @@
     "url": "https://github.com/qiushihe/flowtip/issues"
   },
   "homepage": "https://github.com/qiushihe/flowtip",
-  "dependencies": {
-    "jquery": "^2.1.4",
-    "react": "^0.13.2",
-    "react-tools": "^0.13.2",
-    "requirejs": "^2.1.17",
-    "requirejs-react-jsx": "^0.13.1",
-    "requirejs-text": "^2.0.12",
-    "underscore": "^1.8.3"
+  "scripts": {
+    "dev": "webpack-dev-server --devtool eval --progress --colors --content-base demo --host 0.0.0.0 --history-api-fallback --config webpack.dev.config.babel.js",
+    "build": "webpack -p --config webpack.config.babel.js",
+    "lint": "eslint src"
+  },
+  "peerDepenencies": {
+    "react": "^0.14.3",
+    "react-dom": "^0.14.6"
   },
   "devDependencies": {
-    "amdclean": "^2.7.0",
-    "uglify-js": "^2.4.21"
+    "babel-core": "^6.4.0",
+    "babel-eslint": "^5.0.0-beta6",
+    "babel-loader": "^6.2.1",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.3.13",
+    "eslint": "^1.10.3",
+    "eslint-loader": "^1.2.0",
+    "eslint-plugin-react": "^3.15.0",
+    "path": "^0.12.7",
+    "webpack": "^1.12.11",
+    "webpack-dev-server": "^1.14.1",
+    "webpack-notifier": "^1.2.1"
   }
 }
