--- v1/pages/coordinator-demo.css
@@ -1,16 +1,0 @@
-#demo-container.with-coordinator {
-  position: relative;
-}
-
-.flowtip.with-coordinator {
-  width: 220px;
-  height: 75px;
-  background-color: pink;
-  position: absolute;
-}
-
-.flowtip.with-coordinator .flowtip-tail {
-  position: absolute;
-  width: 20px;
-  height: 10px;
-}
