--- v1/pages/coordinator-demo.html
@@ -1,28 +1,0 @@
-<html>
-<head>
-  <title>FlowTip Coordinator Demo</title>
-  <script src="http://cdnjs.cloudflare.com/ajax/libs/jquery/2.0.2/jquery.min.js"></script>
-  <script src="http://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.4.4/underscore-min.js"></script>
-
-  <script type="text/javascript" src="flowtip.js"></script>
-  <script type="text/javascript" src="coordinator-demo.js"></script>
-  <link rel="stylesheet" media="all" href="demo.css" />
-  <link rel="stylesheet" media="all" href="coordinator-demo.css" />
-</head>
-<body>
-  <h1 id="title">FlowTip Coordinator Demo</h1>
-  <div id="demo-container-wrap">
-    <h3>Container</h3>
-    <div id="demo-container" class="with-coordinator">
-      <div class="flowtip with-coordinator">
-        FlowTip with Coordinator
-        <div class="flowtip-tail"><div></div></div>
-      </div>
-    </div>
-  </div>
-  <div id="demo-code-wrap">
-    <h3>Code</h3>
-    <div id="demo-code"></div>
-  </div>
-</body>
-</html>
