--- v1/pages/demo.html
@@ -1,31 +1,0 @@
-<html>
-<head>
-  <title>FlowTip Demo</title>
-  <script src="http://cdnjs.cloudflare.com/ajax/libs/jquery/2.0.2/jquery.min.js"></script>
-  <script src="http://cdnjs.cloudflare.com/ajax/libs/underscore.js/1.4.4/underscore-min.js"></script>
-
-  <script type="text/javascript" src="flowtip.js"></script>
-  <script type="text/javascript" src="demo.js"></script>
-  <link rel="stylesheet" media="all" href="demo.css" />
-</head>
-<body>
-  <h1 id="title">FlowTip Demo</h1>
-  <ul id="demo-modes">
-    <li><a data-demo-mode="standard">Standard</a></li>
-    <li><a data-demo-mode="persevere">Persevere</a></li>
-    <li><a data-demo-mode="edge-align">Edge Align</a></li>
-    <li><a data-demo-mode="mixed-align">Mixed Align</a></li>
-    <li><a data-demo-mode="target-align">Target Align</a></li>
-    <li><a data-demo-mode="disabled">Disable Regions</a></li>
-    <li><a data-demo-mode="disabled-hidden">Disable Regions (hidden)</a></li>
-  </ul>
-  <div id="demo-container-wrap">
-    <h3>Container</h3>
-    <div id="demo-container"></div>
-  </div>
-  <div id="demo-code-wrap">
-    <h3>Code</h3>
-    <div id="demo-code"></div>
-  </div>
-</body>
-</html>
